Move header test to new file
